Move dataset to its own file
